wavecat: initial commit
This commit is a squash of approximately hundred of patches implementing a new
numerical computation system for special functions of wave catastrophes. These
patches were not so clear, beautiful and bisectabe to be published for external
reader without changes. Unfortunately, now I have to time to make them better,
so, they are simply squashed.

Special functions of wave catastrophes (or SWC) are special functions used
in uniform asymptotics invented in wave catastrophe theory. These asymptotics
allow to compute a field at caustics and at the edge "light-shadow".

A wide range of special functions are supported. All of them can be computed
with respect to a number of parameters, two selected parameters can be changed
in some range with fixed step, others should be constants.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>