core: introduce SCGI mode
This change introduces a new mode of operation - SCGI mode. That means: it is
possible to run WaveCat as SCGI server - an application which can response
requests from HTTP servers with support of SCGI protocol. The implementation is
very dummy and error-prone but that is a very important step towards
clasterization. It also affects performance: in opposite to the CGI mode it
doesn't necessary to start new processes. And the last one, because the process
is alive all the time it is possible to cache computing results without any
backing storage, just in memory. I hope that some kind of memification can help
with performance and give ability to run WaveCat as publicly available service.
Unfortunately, now it is slow and sharing the application via the WEB can be
risky.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>