kernel: introduce result caching subsystem
Computing of SWCs is CPU-intensive task and unfortunately it is not possible to
provide any kind of public service if computing even over a small region of
parameters' values requires seconds of CPU time. Caching of results is a good
approach. First of all it gives not huge overhead, secondly it gives crazy
performance gain, because at some time cache already has necessary values.

This implementation looks for exact matches of numbers. I think that in future
such behavior can be changed a little and search will be available with some
epsilon value as a parameter.

TODO:
	- add synchronization of cache access when parallel computing is
	  enabled;
	- turn global allocation statistics to per-descriptor.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>