kernel: refactor legacy parallel core code a bit
The code of the core was based on a number of very old independent programs for
computing SWCs. Those programs were written by methamatitians (not programmers)
in C with old-stylish Fortran coding approagches. Such code was historically
inherited in the form of catastrophe modules and core functions. Parallel
computing engine was added on top of the core functions. As the result some
strange issues were noticed in core behavior.  Most of the things were rewritten
and now the project has almost no that legacy code, so, some refactoring is
necessary to remove workarounds and strange parts. This commit is a first step
to the cleaner computing kernel.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>