kernel: introduce basic handling for file uploading
Allow the user to upload files with task descriptions instead of wrinting it by
hand in corresponding text area.

To support file uploading some additional handling of the POST request body is
necessary. Now the code assumes that the user can pass only one option: a file
or a hend-written task description, not both of them.

Signed-off-by: Sergei Rogachev <rogachevsergei@gmail.com>